requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
bcrypt==4.1.3
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
dnspython==2.6.1
email_validator==2.1.1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
lxml==5.2.2
lxml_html_clean==0.1.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
parse==1.20.1
passlib==1.7.4
psycopg2==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pyee==11.1.0
Pygments==2.18.0
pyppeteer==2.0.0
pyquery==2.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
requests-html==0.10.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.29.0
w3lib==2.1.2
watchfiles==0.21.0
websockets==12.0
zipp==3.19.1
